requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain-core==0.3.15
google-generativeai==0.8.3
groq>=0.9.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
sse-starlette>=1.6.5,<2.0.0
